pyproject.toml
[tool.poetry]
name = "drawtables"
version = "0.1.0"
description = ""
authors = ["Anton Shushmentsev <shushmencev97@mail.ru>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
uvicorn = "^0.34.2"
fastapi = "^0.115.12"
sqlalchemy = "^2.0.40"
python-dotenv = "^1.1.0"
asyncpg = "^0.30.0"
pydantic = {extras = ["email"], version = "^2.11.4"}
bcrypt = "^4.3.0"
pyjwt = {extras = ["crypto"], version = "^2.10.1"}
python-multipart = "^0.0.20"
redis = "^6.0.0"


[tool.poetry.group.dev.dependencies]
alembic = "^1.15.2"
black = "^25.1.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
